Move file service into a module
Stop being lazy and move the file service abstraction into a
separate module to start a proper service layer.  Separate the
HTTP response handling from the file service and put the latter
in the handler.
